Extract SubmissionButtons into its own file
Move SubmissionButtons and its props interface from rule_form.tsx into
form/components/submission_buttons.tsx for better modularity ahead of
adding the request preview feature.

Made-with: Cursor
