+++ x-pack/platform/packages/shared/response-ops/alerting-v2-rule-form/form/rule_form.tsx
@@ -6,19 +6,12 @@
  */
 
 import React, { useCallback, useRef, useMemo, useState } from 'react';
-import {
-  EuiButton,
-  EuiButtonEmpty,
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiHorizontalRule,
-  EuiSpacer,
-} from '@elastic/eui';
+import { EuiFlexGroup, EuiFlexItem, EuiSpacer } from '@elastic/eui';
 import { useFormContext } from 'react-hook-form';
 import { QueryClient, QueryClientProvider } from '@kbn/react-query';
-import { FormattedMessage } from '@kbn/i18n-react';
 import type { FormValues } from './types';
 import { EditModeToggle, type EditMode } from './components/edit_mode_toggle';
+import { SubmissionButtons } from './components/submission_buttons';
 import {
   RuleFormProvider,
   useRuleFormServices,
@@ -33,7 +26,6 @@ import { NameField } from './fields/name_field';
 import { ErrorCallOut } from './error_callout';
 import { useCreateRule } from './hooks/use_create_rule';
 import { useUpdateRule } from './hooks/use_update_rule';
-import { RULE_FORM_ID } from './constants';
 
 export type { RuleFormServices } from './contexts';
 
@@ -63,59 +55,6 @@ export interface RuleFormProps {
   /** When provided, the form operates in edit mode and uses PATCH instead of POST on submission. */
   ruleId?: string;
 }
-
-interface SubmissionButtonsProps {
-  isSubmitting: boolean;
-  onCancel?: () => void;
-  submitLabel?: React.ReactNode;
-  cancelLabel?: React.ReactNode;
-}
-
-const SubmissionButtons = ({
-  isSubmitting,
-  onCancel,
-  submitLabel,
-  cancelLabel,
-}: SubmissionButtonsProps) => {
-  const defaultSubmitLabel = (
-    <FormattedMessage id="xpack.alertingV2.ruleForm.submitLabel" defaultMessage="Save" />
-  );
-
-  const defaultCancelLabel = (
-    <FormattedMessage id="xpack.alertingV2.ruleForm.cancelLabel" defaultMessage="Cancel" />
-  );
-
-  return (
-    <>
-      <EuiHorizontalRule />
-      <EuiFlexGroup justifyContent="spaceBetween" alignItems="center">
-        {onCancel && (
-          <EuiFlexItem grow={false}>
-            <EuiButtonEmpty
-              onClick={onCancel}
-              isDisabled={isSubmitting}
-              data-test-subj="ruleV2FormCancelButton"
-            >
-              {cancelLabel ?? defaultCancelLabel}
-            </EuiButtonEmpty>
-          </EuiFlexItem>
-        )}
-        <EuiFlexItem grow={false}>
-          <EuiButton
-            type="submit"
-            form={RULE_FORM_ID}
-            isLoading={isSubmitting}
-            fill
-            iconType="plusInCircle"
-            data-test-subj="ruleV2FormSubmitButton"
-          >
-            {submitLabel ?? defaultSubmitLabel}
-          </EuiButton>
-        </EuiFlexItem>
-      </EuiFlexGroup>
-    </>
-  );
-};
 
 /**
  * Inner content component that renders the appropriate form based on edit mode.
